Added Deeplinks and modification in ReminderReceiver Deeplink for create/update route can be triggered by clicking on Open action on notifications and content intent is set to home route deep link. Put NavRoutes NavConstants and other navigation related thing in one package. NextAlarmTimeHelper.kt extra extensions are added to log the correct time. NavHostController is put into MainActivity as MainActivity::onNewIntent requires it to handle deep links.
