+++ app/src/main/java/com/eva/reminders/MainActivity.kt
@@ -1,36 +1,36 @@
 package com.eva.reminders
 
-import android.app.NotificationManager
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Surface
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.core.content.getSystemService
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.eva.reminders.presentation.NavigationGraph
+import androidx.core.view.WindowCompat
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.rememberNavController
+import com.eva.reminders.presentation.navigation.NavigationGraph
+import com.eva.reminders.presentation.utils.LocalSnackBarHostProvider
 import com.eva.reminders.ui.theme.RemindersTheme
-import com.eva.reminders.utils.NotificationConstants
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
 
-    private val notificationManager by lazy { getSystemService<NotificationManager>() }
+    private lateinit var navController: NavHostController
 
     override fun onCreate(savedInstanceState: Bundle?) {
 
+        WindowCompat.setDecorFitsSystemWindows(window, false)
+
         installSplashScreen()
         super.onCreate(savedInstanceState)
-        // If the notification start the activity
-        if (intent.action == NotificationConstants.NOTIFICATION_INTENT_ACTION) {
-            val extra = intent.getIntExtra(NotificationConstants.INTENT_EXTRA_ID, -1)
-            if (extra == -1) return
-            notificationManager?.cancel(extra)
-        }
 
         setContent {
             RemindersTheme {
@@ -38,19 +38,20 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    NavigationGraph()
+                    navController = rememberNavController()
+                    val snackBarHostState = remember { SnackbarHostState() }
+                    CompositionLocalProvider(
+                        LocalSnackBarHostProvider provides snackBarHostState
+                    ) {
+                        NavigationGraph(navHost = navController)
+                    }
                 }
             }
         }
     }
 
     override fun onNewIntent(intent: Intent?) {
-        //If the activity is active and the user clicks the notification.
         super.onNewIntent(intent)
-        if (intent?.action != NotificationConstants.NOTIFICATION_INTENT_ACTION) return
-        // Canceling the requested notification if this is requested via a notification
-        val extra = intent.getIntExtra(NotificationConstants.INTENT_EXTRA_ID, -1)
-        if (extra == -1) return
-        notificationManager?.cancel(extra)
+        navController.handleDeepLink(intent)
     }
 }
