+++ app/src/main/java/com/eva/reminders/presentation/navigation/NavConstants.kt
@@ -1,0 +1,6 @@
+package com.eva.reminders.presentation.navigation
+
+object NavConstants {
+    const val TASK_ID = "taskId"
+    const val TASK_ID_QUERY_PARAMS = "?$TASK_ID={$TASK_ID}"
+}
