+++ app/src/main/java/com/eva/reminders/presentation/navigation/NavRoutes.kt
@@ -1,0 +1,8 @@
+package com.eva.reminders.presentation.navigation
+
+
+sealed class NavRoutes(val route: String) {
+    data object Home : NavRoutes("/")
+    data object AddTask : NavRoutes("/add-task")
+    data object EditLabels : NavRoutes("/edit-labels")
+}
