+++ app/src/main/java/com/eva/reminders/services/ReminderReceiver.kt
@@ -10,6 +10,7 @@ import androidx.core.app.NotificationCompat
 import androidx.core.content.getSystemService
 import com.eva.reminders.MainActivity
 import com.eva.reminders.R
+import com.eva.reminders.presentation.navigation.NavigationDeepLinks
 import com.eva.reminders.utils.IntentsExtra
 import com.eva.reminders.utils.NotificationConstants
 import com.eva.reminders.utils.getFirstSentence
@@ -25,41 +26,64 @@ class ReminderReceiver : BroadcastReceiver() {
         val content = intent.getStringExtra(IntentsExtra.TASK_CONTENT)
         val taskId = intent.getIntExtra(IntentsExtra.TASK_ID, -1)
 
-        val notificationReadRequestCode = -(1 * Random(taskId).nextInt())
-        val activityRequestCode = -(2 * Random(taskId).nextInt())
+        val randomValue = Random(taskId).nextInt()
+
+        val seenPendingIntentRequestCode = -1 * randomValue
+        val homePendingIntentRequestCode = -2 * randomValue
+        val openPendingIntentRequestCode = -3 * randomValue
 
         if (taskId == -1) return
 
         val notificationManager by lazy { context.getSystemService<NotificationManager>() }
 
-        val readIntent = PendingIntent.getBroadcast(
-            context, notificationReadRequestCode,
-            Intent(context, RemoveNotificationReceiver::class.java).apply {
+        val seenNotificationIntent = Intent(context, RemoveNotificationReceiver::class.java)
+            .apply {
                 putExtra(IntentsExtra.TASK_ID, taskId)
-                flags = Intent.FLAG_ACTIVITY_NEW_TASK
+                flags = Intent.FLAG_RECEIVER_REPLACE_PENDING
                 action = NotificationConstants.NOTIFICATION_INTENT_ACTION
-            },
+            }
+
+        val readIntent = PendingIntent.getBroadcast(
+            context,
+            seenPendingIntentRequestCode,
+            seenNotificationIntent,
             PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_IMMUTABLE
         )
 
+        val taskLinkIntent = Intent(context, MainActivity::class.java).apply {
+            data = NavigationDeepLinks.taskUriFromTaskId(taskId)
+            action = Intent.ACTION_VIEW
+        }
+
+        val openTaskPendingIntent = PendingIntent.getActivity(
+            context,
+            openPendingIntentRequestCode,
+            taskLinkIntent,
+            PendingIntent.FLAG_IMMUTABLE
+        )
+
+        val homeLinkIntent = Intent(context, MainActivity::class.java).apply {
+            data = NavigationDeepLinks.homeUri
+            action = Intent.ACTION_VIEW
+        }
+
         val activityIntent = PendingIntent.getActivity(
-            context, activityRequestCode,
-            Intent(context, MainActivity::class.java).apply {
-                putExtra(IntentsExtra.TASK_ID, taskId)
-                action = NotificationConstants.NOTIFICATION_INTENT_ACTION
-            },
+            context,
+            homePendingIntentRequestCode,
+            homeLinkIntent,
             PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
         )
 
-        val readAction =
+        val seenAction =
             NotificationCompat.Action
-                .Builder(
-                    R.drawable.ic_notification_bell,
-                    context.getString(R.string.notification_read_action),
-                    readIntent
-                )
+                .Builder(0, context.getString(R.string.notification_seen_action), readIntent)
                 .setAuthenticationRequired(true)
                 .build()
+
+        val openAction = NotificationCompat.Action
+            .Builder(1, context.getString(R.string.notification_open_action), openTaskPendingIntent)
+            .setAuthenticationRequired(true)
+            .build()
 
 
         val publicNotification =
@@ -78,7 +102,7 @@ class ReminderReceiver : BroadcastReceiver() {
                 .setSmallIcon(R.drawable.ic_notification_bell)
                 .setPriority(NotificationCompat.PRIORITY_HIGH)
                 .setCategory(NotificationCompat.CATEGORY_REMINDER)
-                .setOngoing(false)
+                .setOngoing(true)
                 .setContentTitle(title)
                 .setContentIntent(activityIntent)
                 .apply {
@@ -87,12 +111,13 @@ class ReminderReceiver : BroadcastReceiver() {
                         setStyle(NotificationCompat.BigTextStyle().bigText(content))
                     }
                 }
-                .addAction(readAction)
+                .addAction(seenAction)
+                .addAction(openAction)
                 .setVisibility(NotificationCompat.VISIBILITY_PRIVATE)
                 .setPublicVersion(publicNotification)
                 .setDefaults(Notification.DEFAULT_VIBRATE)
+                .setAutoCancel(true)
                 .build()
-
 
         notificationManager?.notify(taskId, notification)
     }
